requirements.txt
streamlit
google-generativeai
moviepy
imageio-ffmpeg
yt-dlp
youtube-transcript-api